Change perspective camera FOV
This is to improve the sense of scale when using the perspective camera. Otherwise, the perspective camera is too-close (in visual similarity) to the orthographic view.

With the scale of most printers we are dealing with in PrusaSlicer, the viewing distance of one meter seems excessive. (Prusa XL included :P)

Simulating a closer viewing distance provides a perspective that seems closer to a viewer's eys when looking closely at a 3D printer bed. Zooming out, as for larger printers, provides acceptable results. Zooming in provides nearly identical results as before the changes.

In the future (when Prusa XXL is in development?), a setting to change the DefaultDistance based on printer size would be helpful in specifying FOV/focal length/optical properties based on requirements.